requirements.txt
accelerate==1.12.0
aiobotocore==3.1.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aioitertools==0.13.0
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
billiard==4.2.4
botocore==1.42.30
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
datasets==4.5.0
dictdiffer==0.9.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==1.0.0
dvc==3.66.1
dvc-data==3.18.2
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-s3==3.3.0
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.20.3
flatten-dict==0.4.2
flufl.lock==9.0.0
frozenlist==1.8.0
fsspec==2025.10.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
grandalf==0.8
gto==1.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
hydra-core==1.3.2
idna==3.11
iterative-telemetry==0.0.10
Jinja2==3.1.6
jmespath==1.1.0
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
orjson==3.11.7
packaging==26.0
pandas==3.0.0
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruamel.yaml==0.19.1
s3fs==2026.1.0
safetensors==0.7.0
scmrepo==3.6.1
semver==3.0.4
setuptools==80.10.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.2
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.0.0
triton==3.6.0
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
voluptuous==0.16.0
wcwidth==0.5.3
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
zc.lockfile==4.0
